refactor(webhooks): extract WebhookListContent from WebhookListView
